Add cpp file for function defs
